added confirm-dialog to all entities

+++ src/app/event-list/event-list.component.ts
@@ -1,17 +1,27 @@
-import { Component, OnInit } from '@angular/core';
 import {Event} from '../../models/event.model';
 import {EventService} from '../../services/event.service';
-
+import {Component, OnDestroy, OnInit} from '@angular/core';
+import {ConfirmDialogComponent} from '../../@root/confirm-dialog/confirm-dialog.component';
+import {takeUntil} from 'rxjs/operators';
+import {MatDialog} from '@angular/material/dialog';
+import {Subject} from 'rxjs';
 @Component({
   selector: 'app-event-list',
   templateUrl: './event-list.component.html',
   styleUrls: ['./event-list.component.scss']
 })
-export class EventListComponent implements OnInit {
+export class EventListComponent implements OnInit, OnDestroy{
+  // tslint:disable-next-line:variable-name
+  protected _onDestroy = new Subject<void>();
   displayedColumns: string[] = ['id', 'title', 'date', 'location', 'actions'];
   dataSource: Event[] = []; // empty then it would be filled
 
-  constructor(private eventService: EventService) { }
+  constructor(private eventService: EventService, private dialog: MatDialog) { }
+
+  ngOnDestroy(): void {
+    this._onDestroy.next();
+    this._onDestroy.complete();
+  }
 
   ngOnInit(): void {
     this.fetchDataSource();
@@ -23,7 +33,20 @@ export class EventListComponent implements OnInit {
       this.dataSource = data;
     });  }
   onRemoveAccount(id: any): void {
-    this.eventService.removeEventById(id).then(() => this.fetchDataSource());
+   // this.eventService.removeEventById(id).then(() => this.fetchDataSource());
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      hasBackdrop: true,
+      disableClose: false,
+    });
+
+    dialogRef.componentInstance.confirmButtonColor = 'warn';
+
+    dialogRef.afterClosed().pipe(takeUntil(this._onDestroy)).subscribe(isDeleteConfirmed => {
+      console.log('removing: ', isDeleteConfirmed);
+      if (isDeleteConfirmed) {
+        this.eventService.removeEventById(id).then(() => this.fetchDataSource());
+      }
+    });
 
   }
 }
