+++ src/app/publication-list/publication-list.component.ts
@@ -1,20 +1,28 @@
-import { Component, OnInit } from '@angular/core';
+import {Component, OnDestroy, OnInit} from '@angular/core';
 import {Publication} from '../../models/publication.model';
 import {PublicationService} from '../../services/publication.service';
-
+import {ConfirmDialogComponent} from '../../@root/confirm-dialog/confirm-dialog.component';
+import {takeUntil} from 'rxjs/operators';
+import {MatDialog} from '@angular/material/dialog';
+import {Subject} from 'rxjs';
 @Component({
   selector: 'app-publication-list',
   templateUrl: './publication-list.component.html',
   styleUrls: ['./publication-list.component.scss']
 })
-export class PublicationListComponent implements OnInit {
+export class PublicationListComponent implements OnInit, OnDestroy {
+  // tslint:disable-next-line:variable-name
+  protected _onDestroy = new Subject<void>();
   displayedColumns: string[] = ['id', 'titre', 'type', 'dateApparition', 'lien', 'sourcePdf', 'actions'];
   dataSource: Publication[] = []; // empty then it would be filled
-  constructor(private publicationService: PublicationService) { }
+  constructor(private publicationService: PublicationService, private dialog: MatDialog) { }
 
   ngOnInit(): void {
     this.fetchDataSource();
-
+  }
+  ngOnDestroy(): void {
+    this._onDestroy.next();
+    this._onDestroy.complete();
   }
   fetchDataSource(): void {
     this.publicationService.getAllPublications().then(data => {
@@ -22,7 +30,20 @@ export class PublicationListComponent implements OnInit {
     });
   }
   onRemoveAccount(id: any): void {
-    this.publicationService.removePublicationById(id).then(() => this.fetchDataSource());
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      hasBackdrop: true,
+      disableClose: false,
+    });
 
+    dialogRef.componentInstance.confirmButtonColor = 'warn';
+
+    dialogRef.afterClosed().pipe(takeUntil(this._onDestroy)).subscribe(isDeleteConfirmed => {
+      console.log('removing: ', isDeleteConfirmed);
+      if (isDeleteConfirmed) {
+        this.publicationService.removePublicationById(id).then(() => this.fetchDataSource());
+      }
+    });
   }
+
+
 }
