+++ src/app/tool-list/tool-list.component.ts
@@ -1,19 +1,29 @@
-import { Component, OnInit } from '@angular/core';
+import {Component, OnDestroy, OnInit} from '@angular/core';
 import {GLOBAL} from '../app-config';
 import {ToolService} from '../../services/tool.service';
 import {Tool} from '../../models/tool.model';
+import {ConfirmDialogComponent} from '../../@root/confirm-dialog/confirm-dialog.component';
+import {takeUntil} from 'rxjs/operators';
+import {MatDialog} from '@angular/material/dialog';
+import {Subject} from 'rxjs';
 @Component({
   selector: 'app-tool-list',
   templateUrl: './tool-list.component.html',
   styleUrls: ['./tool-list.component.scss']
 })
-export class ToolListComponent implements OnInit {
+export class ToolListComponent implements OnInit, OnDestroy {
+  // tslint:disable-next-line:variable-name
+  protected _onDestroy = new Subject<void>();
   displayedColumns: string[] = ['id', 'date', 'source', 'actions'];
   dataSource: Tool[] = [];
-  constructor(private toolService: ToolService) { }
+  constructor(private toolService: ToolService, private dialog: MatDialog) { }
 
   ngOnInit(): void {
     this.fetchDataSource();
+  }
+  ngOnDestroy(): void {
+    this._onDestroy.next();
+    this._onDestroy.complete();
   }
   fetchDataSource(): void {
     this.toolService.getAllTools().then(data => {
@@ -21,8 +31,23 @@ export class ToolListComponent implements OnInit {
     });  }
 
   onRemoveAccount(id: any): void {
-    this.toolService.removeToolById(id).then(() => this.fetchDataSource());
+    //this.toolService.removeToolById(id).then(() => this.fetchDataSource());
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      hasBackdrop: true,
+      disableClose: false,
+    });
+
+    dialogRef.componentInstance.confirmButtonColor = 'warn';
+
+    dialogRef.afterClosed().pipe(takeUntil(this._onDestroy)).subscribe(isDeleteConfirmed => {
+      console.log('removing: ', isDeleteConfirmed);
+      if (isDeleteConfirmed) {
+        this.toolService.removeToolById(id).then(() => this.fetchDataSource());
+      }
+    });
 
   }
+
+
 
 }
